app/requirements.txt
altair
pandas
streamlit
langchain-core
langchain-azure-ai
azure-ai-inference
langchain
streamlit-authenticator
pyyaml
bcrypt
faiss-cpu
tiktoken
pymupdf
langchain-community
sentence-transformers
numpy<2
yfinance
langchain-mcp-adapters